requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
readability-lxml>=0.8.1
urllib3>=2.0.0

